+++ cdk/bin/app.ts
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import * as cdk from 'aws-cdk-lib';
-// import { SuptAiStack } from '../lib/supt-ai-stack';
+import { SuptAiStack } from '../lib/supt-ai-stack';
 
 const app = new cdk.App();
 
-// new SuptAiStack(app, 'SuptAiStack', {
-//   env: {
-//     account: process.env.CDK_DEFAULT_ACCOUNT,
-//     region: process.env.CDK_DEFAULT_REGION,
-//   },
-// });
+new SuptAiStack(app, 'SuptAiStack', {
+  env: {
+    account: process.env.CDK_DEFAULT_ACCOUNT,
+    region: process.env.CDK_DEFAULT_REGION || 'us-east-1',
+  },
+});